# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Test Project",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Laravel test project server."
        }
    ],
    "paths": {
        "/api/guests/{guestId}": {
            "get": {
                "tags": [
                    "guests.general"
                ],
                "summary": "Get details of a guest.",
                "description": "",
                "operationId": "guests.show",
                "parameters": [
                    {
                        "name": "guestId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuestGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/guests": {
            "post": {
                "tags": [
                    "guests.general"
                ],
                "summary": "Add new user as quest in hotel",
                "description": "",
                "operationId": "guests.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuestGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/{reservationId}": {
            "get": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Get details of a reservation.",
                "description": "",
                "operationId": "reservations.show",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/{reservationId}/room": {
            "get": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Get details of related room.",
                "description": "",
                "operationId": "reservations.rooms.show",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/confirm": {
            "post": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Change the specific reservation's status into confirmed",
                "description": "",
                "operationId": "reservations.confirm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationCONFIRM"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/completed": {
            "post": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Change the specific reservation's status into completed",
                "description": "",
                "operationId": "reservations.completed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationCOMPLETED"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
        "schemas": {
            "GuestGETAttributes": {
                "required": [
                    "id_number",
                    "first_name",
                    "last_name",
                    "email",
                    "phone",
                    "city",
                    "country",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id_number": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "ABS123456"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "example": "john.doe@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "+48710072772"
                    },
                    "country": {
                        "maxLength": 2,
                        "minLength": 2,
                        "example": "PL"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    }
                },
                "type": "object"
            },
            "GuestGETDataItemLinks": {
                "required": [
                    "self"
                ],
                "properties": {
                    "self": {
                        "type": "string",
                        "example": "http://localhost/api/Guests/1"
                    }
                },
                "type": "object"
            },
            "GuestGETDataItem": {
                "required": [
                    "id",
                    "type",
                    "attributes",
                    "links"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Guests"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/GuestGETAttributes"
                    },
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/GuestGETDataItemLinks"
                    }
                },
                "type": "object"
            },
            "GuestGET": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/GuestGETDataItem"
                    }
                },
                "type": "object"
            },
            "GuestPOSTAttributes": {
                "required": [
                    "id_number",
                    "first_name",
                    "last_name",
                    "city",
                    "country"
                ],
                "properties": {
                    "id_number": {
                        "description": "Id number unique",
                        "type": "string",
                        "maxLength": 30,
                        "example": "ACT456999"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Grzegorz"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Grzegorz"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 100,
                        "example": "gmail@gmail.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 30,
                        "example": "+4851051061"
                    },
                    "city": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Warszawa"
                    },
                    "country": {
                        "type": "string",
                        "maxLength": 2,
                        "minLength": 2,
                        "example": "PL"
                    }
                },
                "type": "object"
            },
            "GuestPOSTDataItem": {
                "required": [
                    "type",
                    "attributes"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "guest"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/GuestPOSTAttributes"
                    }
                },
                "type": "object"
            },
            "GuestPOST": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/GuestPOSTDataItem"
                    }
                },
                "type": "object"
            },
            "ReservationCOMPLETEDDataItem": {
                "required": [
                    "type",
                    "id"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "reservation"
                    },
                    "id": {
                        "type": "int",
                        "example": "4"
                    }
                },
                "type": "object"
            },
            "ReservationCOMPLETED": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationCOMPLETEDDataItem"
                    }
                },
                "type": "object"
            },
...
            "ReservationCONFIRMDataItem": {
                "required": [
                    "type",
                    "attributes",
                    "id"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "reservation"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationCONFIRMAttributes"
                    },
                    "id": {
                        "type": "int",
                        "example": "4"
                    }
                },
                "type": "object"
            },
            "ReservationCONFIRM": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationCONFIRMDataItem"
                    }
                },
                "type": "object"
            },
            "ReservationGETAttributes": {
                "required": [
                    "from",
                    "to",
                    "adults_amount",
                    "children_amount",
                    "infants_amount",
                    "first_name",
                    "last_name",
                    "email",
                    "phone",
                    "notes",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "from": {
                        "type": "string",
                        "format": "Y-m-d",
                        "example": "2020-01-01"
                    },
                    "to": {
                        "type": "string",
                        "format": "Y-m-d",
                        "example": "2020-01-03"
                    },
                    "adults_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "2"
                    },
                    "children_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "0"
                    },
                    "infants_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "0"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "example": "john.doe@example.com"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 65535,
                        "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "​pending​",
                            "​confirmed​",
                            "​cancelled​",
                            "​completed​",
                            "​overdue"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    }
                },
                "type": "object"
            },
...
            "ReservationGETRelationships": {
                "required": [
                    "room"
                ],
                "properties": {
                    "room": {
                        "description": "",
                        "type": "object",
                        "$ref": "#/components/schemas/RoomRelatedSingleOutput"
                    }
                },
                "type": "object"
            },
            "ReservationGETDataItem": {
                "required": [
                    "id",
                    "type",
                    "attributes",
                    "relationships",
                    "links"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "reservations"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETAttributes"
                    },
                    "relationships": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETRelationships"
                    },
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETDataItemLinks"
                    }
                },
                "type": "object"
            },
            "ReservationGET": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETDataItem"
                    }
                },
                "type": "object"
            },
...
            "RoomGETDataItem": {
                "required": [
                    "id",
                    "type",
                    "attributes",
                    "links"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "rooms"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETAttributes"
                    },
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETDataItemLinks"
                    }
                },
                "type": "object"
            },
            "RoomGET": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETDataItem"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}